feat(pg-seed): extract seeding utilities into standalone package
- Create new pg-seed package with CSV, JSON, and SQL seeding utilities
- Support both raw pg.Client/PoolClient and wrapper types via ClientInput abstraction
- Use COPY protocol for fast CSV loading via pg-copy-streams
- Update pgsql-test to depend on pg-seed and re-export for backwards compatibility
- Remove csv-parse and pg-copy-streams direct dependencies from pgsql-test
